requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
streamlit==1.25.0
shap==0.42.1
lime==0.2.0.1
optuna==3.4.0
python-dotenv==1.0.0
tqdm==4.66.1

pytest==7.4.0
pytest-cov==4.1.0

black==23.7.0
flake8==6.0.0
mypy==1.5.1 